+++ grpc/clients/application_lifecycle_service_client_test.go
@@ -46,9 +46,10 @@ var _ = Describe("Application Lifecycle Service Client", func() {
 				Name:        "test_container_app",
 				Vendor:      "test_vendor",
 				Description: "test container app",
-				Image:       "http://test.com/container_app_123",
+				Version:     "latest",
 				Cores:       4,
 				Memory:      4096,
+				Source:      "https://path/to/file.zip",
 			})
 		Expect(err).ToNot(HaveOccurred())
 
@@ -61,9 +62,10 @@ var _ = Describe("Application Lifecycle Service Client", func() {
 				Name:        "test_vm_app",
 				Vendor:      "test_vendor",
 				Description: "test vm app",
-				Image:       "http://test.com/vm_app_123",
+				Version:     "latest",
 				Cores:       4,
 				Memory:      4096,
+				Source:      "https://path/to/file.zip",
 			})
 		Expect(err).ToNot(HaveOccurred())
 	})